Move logic for join events in its own listener
